Added covariance_ellipse to localization_util
